Changed the way message count is reset in connections.
Right now, we're using method A, which simply queues up a decrementation
for the message count 30 seconds after sending a message.

If this turns out to not work very well, there's a backup method we can
use:
Foreach connection, create a deque capped at the message limit. When a
message is sent, try to add datetime.now() to the deque.
If there's still room to add messages in the deque, all is fine.
If the deque is full (and it always will be after running for a while)
    take a peek at the first element. If the date we take peek at ism
    ore than 30 seconds old, we pop it, send our message and add
    datetime.now() to the deque as usual.
    Otherwise, the connection is considered completely full.