Changed the variable parsing methods.
It is now possible to include argument variables.
Example usage:
!add command poke $(source:username_raw) just poked $(1)
Note that this does not check if the argument 1 even exists.
TODO: add "options" to, so for example $(1r) would specify that the
argument is required. (and not be printed if it is empty)

You can now pass arguments through to regular variable callbacks.
Example usage:
!add command epm Current $(1) 's per minute is: $(etm;1)
NOTE the semicolon, that's what sets all of this apart.
If an "invalid" argument is passed through to the callbacks, they return None.
If a callback returns None, the MessageAction is cancelled.

To simplify this process, I've had to remove the default key (tb), so if
any command you had did not specify tb as its prefix, now would be a
good time to fix it.