Added a separate table for banphrases
This means we are no longer going to store banphrases in the tb_filters
table.

Next commit will move over old banphrases from the filters table to the
new banphrase table.

Also implemented a warning system, which can be configured in the
settings table for now. Do not rely on those settings though, as I might
change how the settings things work.