Bunch of changes to Banphrases.
There's now a proper admin interface in the admin zone of the bot
website. It can be used to enable/disable banphrases, and it can also be
used to complete remove banphrases. Also, it's of course possible to
create banphrases and edit existing banphrases.

In DB-related news, we removed a foreign key restraint from the added_by
in tb_banphrase. Also added a column for whether the banphrase is
enabled or not.

Made some small modifications to the javascript of the timers page.
This was made to make the code more reusable.