Added a VOD link to the Pleblist History
app.py
Added a template_filter('seconds_to_vodtime')
to convert seconds to VOD Twitch times.

templates/pleblist_history.html
Added a the VOD link to the history

It checks the stream_chunk ended time to avoid errors,
if the stream went down and has more than one stream_chunk.

Tested on a small database, need to see this if it works on the big database.

The link can be wrong if the Streamer is pressing pause during the song or
if the streamvod gets the wrong starting time. (maybe some seconds lag)