Renamed the module from lottery to pointlottery.
Added aliases to the command for start and stop (begin and end).
Fixed the level-check code.
Fixed an error with the "already joined"-code.
Added a status command (!pointlottery status) to print out how many
people are in the current lottery + how many points are in the pot.

TODO: Should we take a tiny fee? 2%? 1%?