Many changes made to the way we store and handle emotes.
Emotes and their data is no longer stored in MySQL, but is instead
stored in Redis.
As much data as possible is also shared between streamers.
A list of emotes and their ID/hash can be found in:
global:emotes:bttv and global:emotes:twitch
Emote data can be found in pajlada:emotes:count,
      pajlada:emotes:epmrecord. replacing pajlada with the streamer name
      of course.

Added a context-method for RedisManager to simplify the usage of
Pipelines.