Linkchecker commands new syntax
Fixed #69

__New Syntax__

!remove link blacklist ID
!remove link whitelist ID
!add link blacklist --shallow google.com
!add link blacklist --deep google.com
Shallow = level 0 is default

If you add a new link it will commit it immediately.

The bot can't add links to the blacklist,
if it finds a bad link it will just cache it but doesn't add it.

Added __Adminlogs__ for the Blacklist and Whitelist
some work for #138

__Known Issue__

User: !add link blacklist 5 4 4
Bot whisper: Successfully added your links

The bot says that he added the links, but it will not add them.
If people use the old syntax, they will not add 0 or 1 to the blacklist.

Since you can add more than one link at the same time you can't set the `bot.whisper` into the `for link in parts`.
It would spam the user with whispers.