Custom raffle message + Hide messages on clr overlay
Boolean to hide on the clr overlay
- Hide duels
- Hide paidtimeouts
- Hide raffles
If the streamer changes the cam to the bottom left and
people don't realy read it on the overlay, sometimes it should be just turned off.

Custom raffle messages, the arguments looks a bit weird.
But at least you can get a custom message with the right times.

It's there tostop the autojoin bots.
It looks so bad for the streamer if there are bots in his channel.
- Chat is slow and out of nowhere there are people who type !join and don't do anything else.
- Or there is a wall of !join at the same time, like spam bots.

Fixed a small bug for python 3.5 where it must be None not 0 :P

Not sure what you changed for usersettings. But i hope this will work with it :D