Download streamer subscriber badge on webapp startup
Version that should be downloaded can be specified in the config.ini
file under the `web` section using the `subscriber_badge_version` key.
Specifying `-1` as the value will disable the subscriber badge
downloading, this can be useful in case you want to use a custom
subscriber badge, or an old one, and don't want it to be overwritten by
this system.