Update contact page with relevant information.
There's now a link to my discord which has pajbot channels for support/questions.

A section has been added to the config file that the /contact page
reads, with the bot maintainer's information. currently the name and
contact_string fields should be filled in. see kkonatestbroadcaster.ini
for an example. This section lets you, the hoster who isn't pajlada, add
contact information on the bot website.

Fixes #652
Fixed #659