Handle different cursor end in Helix Subscriptions endpoint
The Subscriptions endpoint now returns a valid cursor at the end of the
results, and as a result when we query that page we get an empty list of users.
However, the pagination object here does not contain a cursor, so we
check for that as another "exit strategy" now.

Data we got from the last cursor:
{
  "data": [],
  "pagination": {},
}