use `force_verify` parameter on streamer and bot login
This will always prompt user to confirm the authentication, even if they're already accepted it before.
While it adds an extra click every time, it can help in case of debugging the login or when streamer wants to relog after their previous oauth token got invalid (due to a ban) without the need to remove pajbot1 from connections under the Twitch account.