Fix internal commands not showing on the website
Our comparison was bad because internal commands don't have IDs. the ID
is the database ID, and all internal commands have that ID set to None.

This comparison is essentially the old comparison we made before, except
due to the restructure of WebCommand we couldn't make it as simple
(WebCommand no long being derived from Command)