Add basic "Global command cooldown" module support
The module is disabled by default, and commands do not have respect the
module by default even if it is enabled.
For each command you want to use the global command cooldown, you will
need to edit the command and set --use-global-cd or edit the command on
the admin page.